+++ addons/main/magazines/CfgMagAK762New.hpp
@@ -1,87 +1,87 @@
 class hlc_30Rnd_762x39_m_ak_PMAGD;
 class hlc_30Rnd_762x39_MDim_ak_6L10: hlc_30Rnd_762x39_m_ak_6L10 {
     ammo = "HLC_762x39_Tracer_Dim";
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-N-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
+    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-T-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
     displayName = "30Rnd 7.62x39 57-N-231 6L10 Mag (IR-DIM Tracer Mixed)";
     displayNameShort = "FMJ/IR-DIM";
 };
 class hlc_30Rnd_762x39_MDim_ak: hlc_30Rnd_762x39_m_ak {
     ammo = "HLC_762x39_Tracer_Dim";
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-N-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
+    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-T-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
     displayName = "30Rnd 7.62x39 57-N-231 Mag (IR-DIM Tracer Mixed)";
     displayNameShort = "FMJ/IR-DIM";
 };
 class hlc_30Rnd_762x39_MDim_ak_PMAG: hlc_30Rnd_762x39_m_ak_PMAG {
     ammo = "HLC_762x39_Tracer_Dim";
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-N-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
+    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-T-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
     displayName = "30Rnd 7.62x39 57-N-231 PMAG (IR-DIM Tracer Mixed)";
     displayNameShort = "FMJ/IR-DIM";
 };
 class hlc_30Rnd_762x39_MDim_ak_PMAGD: hlc_30Rnd_762x39_m_ak_PMAGD {
     ammo = "HLC_762x39_Tracer_Dim";
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-N-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
+    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-T-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
     displayName = "30Rnd 7.62x39 57-N-231 Tan PMAG (IR-DIM Tracer Mixed)";
     displayNameShort = "FMJ/IR-DIM";
 };
 class hlc_30Rnd_762x39_MDim_ak_Valmet: hlc_30Rnd_762x39_m_ak_Valmet {
     ammo = "HLC_762x39_Tracer_Dim";
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-N-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
+    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-T-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
     displayName = "30Rnd 7.62x39 57-N-231 Valmet Mag (IR-DIM Tracer Mixed)";
     displayNameShort = "FMJ/IR-DIM";
 };
 class hlc_45Rnd_762x39_MDim_rpk: hlc_45Rnd_762x39_m_rpk {
     ammo = "HLC_762x39_Tracer_Dim";
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-N-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 40<br/>Used in: AK-47, RPK";
+    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-T-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 40<br/>Used in: AK-47, RPK";
     displayName = "40Rnd 7.62x39 57-N-231 Bakelite Mag (IR-DIM Tracer Mixed)";
     displayNameShort = "FMJ/IR-DIM";
 };
 class hlc_75Rnd_762x39_MDim_rpk: hlc_75Rnd_762x39_m_rpk {
     ammo = "HLC_762x39_Tracer_Dim";
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-N-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 75<br/>Used in: AK-47, RPK";
+    descriptionShort = "Caliber: 7.62x39mm 57-N-231/57-T-231PMN<br/>Type: FMJ/IR-DIM<br/>Rounds: 75<br/>Used in: AK-47, RPK";
     displayName = "75Rnd 7.62x39 57-N-231 Drum (IR-DIM Tracer Mixed)";
     displayNameShort = "FMJ/IR-DIM";
 };
 
 
 class hlc_30Rnd_762x39_TDim_ak_6L10: hlc_30Rnd_762x39_MDim_ak_6L10 {
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231PMN<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
-    displayName = "30Rnd 7.62x39 57-N-231PMN 6L10 Mag (IR-DIM Tracer)";
+    descriptionShort = "Caliber: 7.62x39mm 57-T-231PMN<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
+    displayName = "30Rnd 7.62x39 57-T-231PMN 6L10 Mag (IR-DIM Tracer)";
     displayNameShort = "IR-DIM";
     tracersEvery = 1;
 };
 class hlc_30Rnd_762x39_TDim_ak: hlc_30Rnd_762x39_MDim_ak {
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231PMN<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
-    displayName = "30Rnd 7.62x39 57-N-231PMN Mag (IR-DIM Tracer)";
+    descriptionShort = "Caliber: 7.62x39mm 57-T-231PMN<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
+    displayName = "30Rnd 7.62x39 57-T-231PMN Mag (IR-DIM Tracer)";
     displayNameShort = "IR-DIM";
     tracersEvery = 1;
 };
 class hlc_30Rnd_762x39_TDim_ak_PMAG: hlc_30Rnd_762x39_MDim_ak_PMAG {
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231PMN<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
-    displayName = "30Rnd 7.62x39 57-N-231PMN Black PMAG (IR-DIM Tracer)";
+    descriptionShort = "Caliber: 7.62x39mm 57-T-231PMN<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
+    displayName = "30Rnd 7.62x39 57-T-231PMN PMAG (IR-DIM Tracer)";
     displayNameShort = "IR-DIM";
     tracersEvery = 1;
 };
 class hlc_30Rnd_762x39_TDim_ak_PMAGD: hlc_30Rnd_762x39_MDim_ak_PMAGD {
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231PMN<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
-    displayName = "30Rnd 7.62x39 57-N-231PMN Tan PMAG (IR-DIM Tracer)";
+    descriptionShort = "Caliber: 7.62x39mm 57-T-231PMN<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
+    displayName = "30Rnd 7.62x39 57-T-231PMN Tan PMAG (IR-DIM Tracer)";
     displayNameShort = "IR-DIM";
     tracersEvery = 1;
 };
 class hlc_30Rnd_762x39_TDim_ak_Valmet: hlc_30Rnd_762x39_MDim_ak_Valmet {
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231PMN<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
-    displayName = "30Rnd 7.62x39 57-N-231PMN Valmet Mag (IR-DIM Tracer)";
+    descriptionShort = "Caliber: 7.62x39mm 57-T-231PMN<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-47, RPK";
+    displayName = "30Rnd 7.62x39 57-T-231PMN Valmet Mag (IR-DIM Tracer)";
     displayNameShort = "IR-DIM";
     tracersEvery = 1;
 };
 class hlc_45Rnd_762x39_TDim_rpk: hlc_45Rnd_762x39_MDim_rpk {
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231PMN<br/>Type: IR-DIM<br/>Rounds: 40<br/>Used in: AK-47, RPK";
-    displayName = "40Rnd 7.62x39 57-N-231PMN Bakelite Mag (IR-DIM Tracer)";
+    descriptionShort = "Caliber: 7.62x39mm 57-T-231PMN<br/>Type: IR-DIM<br/>Rounds: 40<br/>Used in: AK-47, RPK";
+    displayName = "40Rnd 7.62x39 57-T-231PMN Bakelite Mag (IR-DIM Tracer)";
     displayNameShort = "IR-DIM";
     tracersEvery = 1;
 };
 class hlc_75Rnd_762x39_TDim_rpk: hlc_75Rnd_762x39_MDim_rpk {
-    descriptionShort = "Caliber: 7.62x39mm 57-N-231PMN<br/>Type: IR-DIM<br/>Rounds: 75<br/>Used in: AK-47, RPK";
-    displayName = "75Rnd 7.62x39 57-N-231PMN Drum (IR-DIM Tracer)";
+    descriptionShort = "Caliber: 7.62x39mm 57-T-231PMN<br/>Type: IR-DIM<br/>Rounds: 75<br/>Used in: AK-47, RPK";
+    displayName = "75Rnd 7.62x39 57-T-231PMN Drum (IR-DIM Tracer)";
     displayNameShort = "IR-DIM";
     tracersEvery = 1;
 };